Move ShowRouteInfoDialogFragment to routepreparationmenu
